Renamed "macros" to "html_macro"
